ci: make github action pure
