+++ README.md
@@ -6,11 +6,12 @@ Repoistory to host all of Fuel's reusable workflows
 
 | Group                             | Description                                                      |
 | --------------------------------- | ---------------------------------------------------------------- |
+| [changeset](./changeset/)         | Reusable workflow for create changesets and release npm packages |
 | [gh-projects](./gh-projects/)     | Automating interactions between GH Projects and repositories     |
 | [setups/node](./setups/node/)     | Setup node and pnpm requirements on CI                           |
 | [setups/docker](./setups/docker/) | Setup docker and docker compose on CI                            |
 | [setups/npm](./setups/npm/)       | Setup npm deployment requirements on CI                          |
-| [changeset](./changeset/)         | Reusable workflow for create changesets and release npm packages |
+| [update-sdk](./update-sdk/)       | Reusable workflow for update the SDK packages                    |
 
 ## License
