--- gh-projects/update-sdk/src/process.env.d.ts
@@ -1,12 +1,0 @@
-declare global {
-  namespace NodeJS {
-    interface ProcessEnv {
-      GITHUB_TOKEN: string;
-      GITHUB_ORGANIZATION: string;
-      GITHUB_REPOSITORY: string;
-      NPM_TAG: 'next' | 'latest';
-    }
-  }
-}
-
-export {};
