+++ update-sdk/package.json
@@ -5,6 +5,7 @@
   "main": "./src/index.ts",
   "scripts": {
     "ts:check": "tsc --noEmit",
+    "build": "tsup",
     "check-versions": "tsx src/index.ts"
   },
   "devDependencies": {
@@ -12,6 +13,10 @@
     "@types/node": "20.11.29",
     "chalk": "5.3.0",
     "execa": "8.0.1",
+    "tsup": "8.0.2",
     "tsx": "4.7.1"
+  },
+  "dependencies": {
+    "@actions/core": "1.10.1"
   }
 }
