+++ update-sdk/src/PackageJson.ts
@@ -3,7 +3,7 @@ import { exec } from 'node:child_process';
 import { promisify } from 'node:util';
 
 export class PackageJson {
-  static async updateDependencies(tag: 'next' | 'latest') {
+  static async updateDependencies(tag: string) {
     console.log(c.white(`📟 Running pnpm update with tag ${tag}...\n`));
     await promisify(exec)(
       `pnpm update "@fuels/react@${tag}" "@fuels/connectors@${tag}" "fuels@${tag}" --recursive`
